Crate gan1_beta1

This documentation was generated from gan crate version 0.0.1. The original source code can be found on github.

Features

Handle the following Resources with ease from the central hub ...

advertisers (get and list)

cc offers (list)

events (list)

links (get, insert and list)

publishers (get and list)

reports (get)

Everything else about the gan v1_beta1 API can be found at the official documentation site.

Not what you are looking for ? Find all other google APIs in their Rust documentation index.

Structure of this Library

The API is structured into the following primary items:

Hub

a central object to maintain state and allow accessing all Activities

Resources

primary types that you can apply Activities to

a collection of properties and Parts

Parts

a collection of properties

never directly used in Activities

Activities

operations to apply to Resources

Generally speaking, you can invoke Activities like this:

let r = hub.resource().activity(...).doit()

Or specifically ...

let r = hub.links().insert(...).doit()
let r = hub.links().get(...).doit()
let r = hub.links().list(...).doit()

The resource() and activity(...) calls create builders. The second one dealing with Activities supports various methods to configure the impending operation (not shown here). It is made such that all required arguments have to be specified right away (i.e. (...)), whereas all optional ones can be build up as desired. The doit() method performs the actual communication with the server and returns the respective result.

Usage

Setting up your Project

To use this library, you would put the following lines into your Cargo.toml file:

[dependencies]
gan1_beta1 = "0.0.1"

A complete example

extern crate hyper;
extern crate "yup-oauth2" as oauth2;
extern crate "rustc-serialize" as rustc_serialize;
extern crate gan1_beta1;
use gan1_beta1::cmn::Result;
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use gan1_beta1::Gan;
 
// Get an ApplicationSecret instance by some means. It contains the `client_id` and 
// `client_secret`, among other things.
let secret: ApplicationSecret = Default::default();
// Instantiate the authenticator. It will choose a suitable authentication flow for you, 
// unless you replace  `None` with the desired Flow.
// Provide your own `AuthenticatorDelegate` to adjust the way it operates and get feedback about 
// what's going on. You probably want to bring in your own `TokenStorage` to persist tokens and
// retrieve them from storage.
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Gan::new(hyper::Client::new(), auth);
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.links().list("role", "roleId")
             .start_date_min("sadipscing")
             .start_date_max("aliquyam")
             .search_text("ea")
             .relationship_status("no")
             .add_promotion_type("justo")
             .page_token("justo")
             .max_results(67)
             .link_type("et")
             .create_date_min("diam")
             .create_date_max("ipsum")
             .authorship("Lorem")
             .add_asset_size("et")
             .add_advertiser_id("duo")
             .doit();
 
match result {
    Result::HttpError(err) => println!("HTTPERROR: {:?}", err),
    Result::MissingAPIKey => println!("Missing API Key"),
    Result::MissingToken => println!("Missing Token"),
    Result::Failure(_) => println!("General Failure (Response doesn't print)"),
    Result::FieldClash(clashed_field) => println!("FIELD CLASH: {:?}", clashed_field),
    Result::Success(_) => println!("Success (value doesn't print)"),
}

Handling Errors

All errors produced by the system are provided either as Result enumeration as return value of the doit() methods, or handed as possibly intermediate results to either the Hub Delegate, or the Authenticator Delegate.

When delegates handle errors or intermediate values, they may have a chance to instruct the system to retry. This makes the system potentially resilient to all kinds of errors.

About Customization/Callbacks

You may alter the way an doit() method is called by providing a delegate to the Method Builder before making the final doit() call. Respective methods will be called to provide progress information, as well as determine whether the system should retry on failure.

The delegate trait is default-implemented, allowing you to customize it with minimal effort.

About Parts

All structures provided by this library are made to be enocodable and decodable via json. Optionals are used to indicate that partial requests are responses are valid. Most optionals are are considered Parts which are identifyable by name, which will be sent to the server to indicate either the set parts of the request or the desired parts in the response.

About Builder Arguments

Using method builders, you are able to prepare an action call by repeatedly calling it's methods. These will always take a single argument, for which the following statements are true.

PODs are handed by copy

strings are passed as &str

request values are borrowed

Arguments will always be copied or cloned into the builder, to make them independent of their original life times.

Modules

cmn

Macros

map!
This macro is advertised in the documentation, which is why we deliver it as well

Structs

Advertiser
An AdvertiserResource.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list advertisers (none)

get advertisers (response)

AdvertiserGetMethodBuilder
Retrieves data about a single advertiser if that the requesting advertiser/publisher has access to it. Only publishers can lookup advertisers. Advertisers can request information about themselves by omitting the advertiserId query parameter.

AdvertiserListMethodBuilder
Retrieves data about all advertisers that the requesting advertiser/publisher has access to.

AdvertiserMethodsBuilder
A builder providing access to all methods supported on advertiser resources. It is not used directly, but through the Gan hub.

Advertisers
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list advertisers (response)

CcOffer
A credit card offer. There are many possible result fields. We provide two different views of the data, or "projections." The "full" projection includes every result field. And the "summary" projection, which is the default, includes a smaller subset of the fields. The fields included in the summary projection are marked as such in their descriptions.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list cc offers (none)

CcOfferBonusRewards
For cards with rewards programs, extra circumstances whereby additional rewards may be granted.

This type is not used in any activity, and only used as part of another schema.

CcOfferDefaultFees
Fees for defaulting on your payments.

This type is not used in any activity, and only used as part of another schema.

CcOfferListMethodBuilder
Retrieves credit card offers for the given publisher.

CcOfferMethodsBuilder
A builder providing access to all methods supported on ccOffer resources. It is not used directly, but through the Gan hub.

CcOfferRewards
For cards with rewards programs, detailed rules about how the program works.

This type is not used in any activity, and only used as part of another schema.

CcOffers
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list cc offers (response)

Event
An EventResource.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list events (none)

EventListMethodBuilder
Retrieves event data for a given advertiser/publisher.

EventMethodsBuilder
A builder providing access to all methods supported on event resources. It is not used directly, but through the Gan hub.

EventProducts
Products associated with the event.

This type is not used in any activity, and only used as part of another schema.

Events
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list events (response)

Gan
Central instance to access all Gan related resource activities

Link
A LinkResource.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

insert links (request|response)

get links (response)

list links (none)

LinkGetMethodBuilder
Retrieves data about a single link if the requesting advertiser/publisher has access to it. Advertisers can look up their own links. Publishers can look up visible links or links belonging to advertisers they are in a relationship with.

LinkInsertMethodBuilder
Inserts a new link.

LinkListMethodBuilder
Retrieves all links that match the query parameters.

LinkMethodsBuilder
A builder providing access to all methods supported on link resources. It is not used directly, but through the Gan hub.

LinkSpecialOffers
Special offers on the link.

This type is not used in any activity, and only used as part of another schema.

Links
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list links (response)

Money
An ApiMoneyProto.

This type is not used in any activity, and only used as part of another schema.

Publisher
A PublisherResource.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list publishers (none)

get publishers (response)

PublisherGetMethodBuilder
Retrieves data about a single advertiser if that the requesting advertiser/publisher has access to it. Only advertisers can look up publishers. Publishers can request information about themselves by omitting the publisherId query parameter.

PublisherListMethodBuilder
Retrieves data about all publishers that the requesting advertiser/publisher has access to.

PublisherMethodsBuilder
A builder providing access to all methods supported on publisher resources. It is not used directly, but through the Gan hub.

Publishers
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list publishers (response)

Report
A ReportResource representing a report of a certain type either for an advertiser or publisher.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

get reports (response)

ReportGetMethodBuilder
Retrieves a report of the specified type.

ReportMethodsBuilder
A builder providing access to all methods supported on report resources. It is not used directly, but through the Gan hub.